v1.2.5

+++ src/ui_general_screen.cpp
@@ -46,7 +46,7 @@ void createGeneralScreen() {
     lv_obj_set_style_radius(sw_lyrics, 13, LV_PART_MAIN);
     lv_obj_set_style_bg_color(sw_lyrics, lv_color_hex(0x333333), LV_PART_MAIN);
     // Indicator (colored fill when checked) - no padding so it fills the track
-    lv_obj_set_style_bg_color(sw_lyrics, COL_ACCENT, (lv_style_selector_t)(LV_PART_INDICATOR | LV_STATE_CHECKED));
+    lv_obj_set_style_bg_color(sw_lyrics, COL_ACCENT, (lv_style_selector_t)((uint32_t)LV_PART_INDICATOR | (uint32_t)LV_STATE_CHECKED));
     lv_obj_set_style_radius(sw_lyrics, 13, LV_PART_INDICATOR);
     lv_obj_set_style_pad_all(sw_lyrics, 0, LV_PART_INDICATOR);
     // Knob (circle)
